+++ .changeset/stale-panthers-shake.md
@@ -1,0 +1,5 @@
+---
+"@changesets/types": major
+---
+
+Properly sort this into public and private types, and rename types for actual usage
